First refactor pass for create_webdataset.py
