+++ src/dataset_tools/create_webdataset.py
@@ -15,10 +15,10 @@ import webdataset as wds
 from PIL import Image
 from torchvision import transforms
 
-from src.dataset_tools.utils import set_random_seeds
+from src.dataset_tools.utils import set_random_seeds, square_crop
 
 
-def prepare_json_data(sample_metadata, columns_to_json):
+def _prepare_json_data(sample_metadata, columns_to_json):
     metadata = {}
     columns_to_json = columns_to_json.split(",")
 
@@ -28,46 +28,7 @@ def prepare_json_data(sample_metadata, columns_to_json):
     return metadata
 
 
-def square_crop(image, x, y, width, height):
-    """
-    Try to crop a square region centered at the bounding box provided. The square
-    region cropped has a padding corresponding to 10% of the side of a tight
-    square cropping around the bbox. If the square region would crop outside the
-    image size, the region will be capped to be at the borders. In this case, the
-    region is not guaranteed to be square.
-
-    Args:
-      image: a PIL image
-      x (float): left coordinate
-      y (float): upper coordinate
-      width (float): bbox width
-      height (float): bbox height
-
-    Returns:
-      image: a cropped image
-    """
-
-    image_width, image_height = image.size
-    x = x * image_width
-    y = y * image_height
-    width = width * image_width
-    height = height * image_height
-
-    side = 1.2 * max(width, height)
-    center_x = x + width / 2
-    center_y = y + height / 2
-
-    upper = max(0.0, center_y - side / 2)
-    left = max(0.0, center_x - side / 2)
-    lower = min(image_height, center_y + side / 2)
-    right = min(image_width, center_x + side / 2)
-
-    image = image.crop((int(left), int(upper), int(right), int(lower)))
-
-    return image
-
-
-def crop_to_bbox(image, md_preds):
+def _crop_to_bbox(image, md_preds):
     if len(md_preds["detections"]) > 0:
         max_detection_conf = 0.0
         (x, y, width, height) = 0, 0, 0, 0
@@ -83,13 +44,13 @@ def crop_to_bbox(image, md_preds):
     return image
 
 
-def get_resize_transform(resize_min_size):
+def _get_resize_transform(resize_min_size):
     return transforms.Compose(
         [transforms.Resize(resize_min_size), transforms.ToTensor()]
     )
 
 
-def get_category_map(
+def _get_category_map(
     dataset_df, label_column: str, category_map_json: str, save_category_map_json: str
 ):
     if category_map_json is not None:
@@ -106,7 +67,8 @@ def get_category_map(
     return categories_map
 
 
-def dataset_samples(
+# TODO : Refactor _dataset_sample
+def _dataset_samples(
     annotations_csv: str,
     dataset_dir: str,
     shuffle_images: bool,
@@ -129,12 +91,12 @@ def dataset_samples(
     dataset_df = pd.read_csv(annotations_csv)
     if shuffle_images:
         dataset_df = dataset_df.sample(frac=1)
-    categories_map = get_category_map(
+    categories_map = _get_category_map(
         dataset_df, label_column, category_map_json, save_category_map_json
     )
 
     if resize_min_size is not None:
-        resize_transform = get_resize_transform(resize_min_size)
+        resize_transform = _get_resize_transform(resize_min_size)
 
     for _, row in dataset_df.iterrows():
         fpath = os.path.join(dataset_dir, row[image_path_column])
@@ -156,7 +118,7 @@ def dataset_samples(
             if md_results is not None:
                 try:
                     md_preds = md_results.loc[row[image_path_column]]
-                    image = crop_to_bbox(image, md_preds)
+                    image = _crop_to_bbox(image, md_preds)
                 except KeyError:
                     print("Skipping crop for the image: ", row[image_path_column])
                 except TypeError:
@@ -179,7 +141,7 @@ def dataset_samples(
         }
 
         if columns_to_json is not None:
-            json_data = prepare_json_data(row, columns_to_json)
+            json_data = _prepare_json_data(row, columns_to_json)
             sample["json"] = json_data
 
         yield sample
@@ -202,7 +164,7 @@ def create_webdataset(
 ):
     set_random_seeds(random_seed)
     with wds.ShardWriter(webdataset_patern, maxsize=max_shard_size) as sink:
-        dataset = dataset_samples(
+        dataset = _dataset_samples(
             annotations_csv,
             dataset_dir,
             shuffle_images,
